BookController and ReaderController implementation

+++ src/main/java/com/restapilibrary/service/BookService.java
@@ -1,50 +1,51 @@
 package com.restapilibrary.service;
 
 import com.restapilibrary.domain.Book;
+import com.restapilibrary.exceptions.BookNotFoundException;
 import com.restapilibrary.repository.BookRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.context.event.ApplicationReadyEvent;
-import org.springframework.context.event.EventListener;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDate;
 import java.util.List;
-import java.util.Optional;
 
 @Service
+@RequiredArgsConstructor
 public class BookService {
 
-    private BookRepository bookRepository;
-
-    @Autowired
-    public BookService(BookRepository bookRepository) {
-        this.bookRepository = bookRepository;
-    }
-
-
-    public Optional<Book> findBookById(Long bookId) {
-        return bookRepository.findById(bookId);
-    }
+    private final BookRepository bookRepository;
 
     public List<Book> findAllBooks() {
         return bookRepository.findAll();
+    }
+
+    public Book findBookById(final Long bookId) throws BookNotFoundException {
+        return bookRepository.findById(bookId).orElseThrow(BookNotFoundException::new);
     }
 
     public Book saveBook(Book book) {
         return bookRepository.save(book);
     }
 
-    public void deleteBook(Long bookId) {
-        bookRepository.deleteById(bookId);
+    public Book updateBook(final Book book) throws BookNotFoundException {
+        if (bookRepository.existsById(book.getBookId())) {
+            bookRepository.save(book);
+            return book;
+        } else {
+            throw new BookNotFoundException();
+        }
     }
 
-    @EventListener(ApplicationReadyEvent.class)
-    public void fillDb() {
-        saveBook(new Book(1L, "Potop", "Henryk Sienkiewicz", LocalDate.of(1880, 1,1)));
-        saveBook(new Book(2L, "Król", "Szczepan Twardoch", LocalDate.of(2018, 2,22)));
+    public void deleteBook(Long bookId) throws BookNotFoundException {
+        if (bookRepository.existsById(bookId)) {
+            bookRepository.deleteById(bookId);
+        } else {
+            throw new BookNotFoundException();
+        }
     }
 
-
-
-
+//    @EventListener(ApplicationReadyEvent.class)
+//    public void fillDb() {
+//        saveBook(new Book(1L, "Potop", "Henryk Sienkiewicz", LocalDate.of(1880, 1,1)));
+//        saveBook(new Book(2L, "Król", "Szczepan Twardoch", LocalDate.of(2018, 2,22)));
+//    }
 }
